+++ changelog.md
@@ -1,6 +1,22 @@
 # Change Log for PSHyperV
 
+## v0.2.0
+
++ Added `mygroupvm.format.ps1xml`
++ Updated `README.md`
++ Added `Find-VMGroup` which is a proxy version of `Get-VMGroup`
++ Added functions:
+  + `Get-VMIPAddress`
+  + `Open-VMRemoteDesktop`
+  + `Open-VMConnect`
+  + `Get-VHDSummary`
+  + `Get-VMState`
+  + `Get-VMMemorySummary`
++ Added `vmIPAddress.format.ps1xml`
++ Added `vmMemorySummary.format.ps1xml`
++ Added help markdown files
++ Added external help
+
 ## v0.1.0
 
 + initial files
-